chore(autoresearch): retire AutoScrape + AutoQuality harnesses
Stream 7 of the scraping rethink found AutoScrape produced zero outputs
ever (the Pictures/AutoResearch/AutoScrape/ directory in Obsidian where
it was supposed to write reports doesn't exist). The BFI IMAX silent-
breaker case on 2026-04-27/28 — the textbook scenario the harness was
built for — was missed entirely.

AutoQuality plateaued at DQS 90.0-90.3 across 30 experiments without
finding any beneficial threshold change. Its own report concluded the
biggest remaining lever is direct enrichment via Claude Code scripts —
which is what we now do.

Sunk-cost recovery (Stream 8): ~£8.70/month in Gemini 3.1 Pro API calls.

- Delete src/autoresearch/ entirely (16 files: autoscrape/*, autoquality/*,
  experiment-log.ts, obsidian-reporter.ts, types.ts).
- Delete .claude/commands/{autoscrape,autoquality}.md (local).
- Relocate the thresholds config that DOES still drive the data-quality
  pipeline (consumed by tmdb/match.ts + 3 other call sites):
    src/autoresearch/autoquality/thresholds.json
      → src/lib/data-quality/thresholds.json
    src/autoresearch/autoquality/load-thresholds.ts
      → src/lib/data-quality/load-thresholds.ts (rewritten — drops DB
        loading path and cache-mutation surface; thresholds are now a
        static bundled JSON, edit the file to change them).
- Update 4 importers to the new path.

Replacement for AutoScrape's job: src/lib/scrape-quarantine.ts
(Prowlarr-style health-state-machine, shipped in the prior commit).

Tests: 918/918. Typecheck clean.
